requirements.txt
flask
python-telegram-bot==13.15
psycopg2-binary
scikit-learn
xgboost
numpy
joblib
tensorflow
